requirements.txt
openai==2.7.2
Pillow==12.0.0
selenium==4.38.0
streamlit==1.51.0
openpyxl==3.1.5
xlrd==2.0.1
olefile==0.47
msoffcrypto-tool==5.4.2
